split out tests and fix lint
